ICU-20593 Renaming Python buildtool to icutools.databuilder.
