+++ icu4c/source/configure.ac
@@ -1399,7 +1399,8 @@ else
     if test -f "$srcdir/data/locales/root.txt";
     then
         echo "Spawning Python to generate data/rules.mk..."
-        PYTHONPATH="$srcdir/data" $PYTHON -m buildtool \
+        PYTHONPATH="$srcdir/python" \
+            $PYTHON -m icutools.databuilder \
             --mode gnumake \
             --seqmode parallel \
             --src_dir "$srcdir/data" \
@@ -1414,7 +1415,8 @@ else
       touch data/rules.mk
     fi
     echo "Spawning Python to generate test/testdata/rules.mk..."
-    PYTHONPATH="$srcdir/test/testdata:$srcdir/data" $PYTHON -m buildtool \
+    PYTHONPATH="$srcdir/python" \
+        $PYTHON -m icutools.databuilder \
         --mode gnumake \
         --seqmode parallel \
         --src_dir "$srcdir/test/testdata" \
