+++ icu4c/source/data/makedata.mak
@@ -233,7 +233,8 @@ $(TOOLS_TS): "$(ICUTOOLS)\genrb\$(CFGTOOLS)\genrb.exe" "$(ICUTOOLS)\gencnval\$(C
 
 $(COREDATA_TS):
 	@cd "$(ICUSRCDATA)"
-	py -3 -B -m buildtool \
+	set PYTHONPATH=$(ICUP)\source\python;%PYTHONPATH%
+	py -3 -B -m icutools.databuilder \
 		--mode windows-exec \
 		--src_dir "$(ICUSRCDATA)" \
 		--tool_dir "$(ICUTOOLS)" \
