+++ icu4c/source/python/icutools/databuilder/comment_stripper.py
