+++ icu4c/source/python/icutools/databuilder/locale_dependencies.py
