+++ icu4c/source/test/testdata/BUILDRULES.py
@@ -3,8 +3,8 @@
 
 from distutils.sysconfig import parse_makefile
 
-from buildtool import *
-from buildtool.request_types import *
+from icutools.databuilder import *
+from icutools.databuilder.request_types import *
 
 
 def generate(config, glob, common_vars):
