+++ icu4c/source/test/testdata/testdata.mak
@@ -34,8 +34,8 @@ CREATE_DIRS :
 
 "$(TESTDATAOUT)\testdata.dat" :
 	@echo Building test data
-	set PYTHONPATH=$(ICUSRCDATA);%PYTHONPATH%
-	py -3 -B -m buildtool \
+	set PYTHONPATH=$(ICUP)\source\python;%PYTHONPATH%
+	py -3 -B -m icutools.databuilder \
 		--mode windows-exec \
 		--tool_dir "$(ICUTOOLS)" \
 		--tool_cfg "$(CFG)" \
